Now with working doubles.

+++ src/camera.cpp
@@ -1,8 +1,8 @@
 #include "camera.h"
 
-Camera::Camera(glm::vec3 pos, int fov, float asp, int focal) {
+Camera::Camera(glm::dvec3 pos, int fov, double asp, int focal) {
     m_pos = pos;
-    m_fov = glm::radians((float) fov);
+    m_fov = glm::radians((double) fov);
     m_asp = asp;
     m_focal = focal;
     m_height = floor(2*m_focal*glm::tan(m_fov/2));
@@ -27,7 +27,7 @@ int Camera::getWidth() {
     return ((int) m_width);
 }
 
-glm::vec3 Camera::getPosition() { 
+glm::dvec3 Camera::getPosition() { 
     return m_pos;
 }
